refactor: extract chat component and transport
